improved Safe-Mode

+++ commands.py
@@ -67,16 +67,18 @@ try:
     answer = 0
     ssh = False
 
-    def CommandAsk(plt=0, USNAME_PRINT=0):
+    def CommandAsk(plt=0, USNAME_PRINT=0, safe_mode=False):
         if MD == "2":
             CommandList(Command=input(f"!History isn't enabled! PyTerminal Beta | {USNAME_PRINT.capitalize()} % "), cmd_pl=plt)
         elif MD == "9": 
             CommandList(Command=input(f"PyTerminal {sys_detect.processor} | {sys_detect.system} {sys_detect.machine} % "), cmd_pl=plt)
+        elif MD == "3":
+            CommandList(Command=input(f"PyTerminal | Safe-Mode $ "), cmd_pl=plt, safe_md=safe_mode)
         else:
             CommandList(Command=input(f"PyTerminal Beta | {USNAME_PRINT.capitalize()} $ "), cmd_pl=plt) 
 
 
-    def CommandList(Command=0, cmd_pl=0):
+    def CommandList(Command=0, cmd_pl=0, safe_md=False):
         global jump, jump_user, ask_recv, LCommand, answer
         LCommand = 0
 
@@ -122,45 +124,46 @@ try:
             CommandSay(answer=now.strftime("%Y-%m-%d %H:%M:%S"))
 
         if Command == "del" or Command == "delete":
-            
-            if MD == "2":
-                CommandSay(answer=os.listdir(dir))
-                ask_del = input("what file you want to delete:")
-                try:
-                    CommandSay(answer=ask_del)
-                    os.remove(ask_del)
-                    CommandSay(answer="DONE", color="OKGREEN")
-                except FileNotFoundError:
-                    CommandSay(answer="This file doesn't exist", color="FAIL")
-            else:
-                LCommand = Command
-                CommandSay(answer="This Function isn't available within this mode", color="FAIL")
+            if not safe_md:
+                if MD == "2":
+                    CommandSay(answer=os.listdir(dir))
+                    ask_del = input("what file you want to delete:")
+                    try:
+                        CommandSay(answer=ask_del)
+                        os.remove(ask_del)
+                        CommandSay(answer="DONE", color="OKGREEN")
+                    except FileNotFoundError:
+                        CommandSay(answer="This file doesn't exist", color="FAIL")
+                else:
+                    LCommand = Command
+                    CommandSay(answer="This Function isn't available within this mode", color="FAIL")
 
         if Command == "create":
-            if MD == "1":
-                LCommand = Command
-                ask_name = input("What the name of the file you want to create?")
-                try:
-                    open(ask_name, "x")
-                    CommandSay(answer="DONE", color="OKGREEN")
-                except FileExistsError:
-                    ask_del_create = input("This file already exist try again", color="WARNING")
-                except UnboundLocalError:
-                    CommandSay(answer="There was a Problem try again", color="FAIL")
-            elif MD == "2" or MD == "9":
+            if not safe_md:
+                if MD == "1":
+                    LCommand = Command
                     ask_name = input("What the name of the file you want to create?")
                     try:
                         open(ask_name, "x")
                         CommandSay(answer="DONE", color="OKGREEN")
                     except FileExistsError:
-                        ask_del_create = input("This file already exist do you want to delete it. if yes type 'Y'")
-                        if ask_del_create == "Y" or ask_del_create == "y":
-                            os.remove(ask_name)
-                            CommandSay(answer="DONE", color="OKGREEN")
+                        ask_del_create = input("This file already exist try again", color="WARNING")
                     except UnboundLocalError:
-                            pass
-            else:
-                CommandSay(answer="This Function isn't available within this mode", color="FALI")
+                        CommandSay(answer="There was a Problem try again", color="FAIL")
+                elif MD == "2" or MD == "9":
+                        ask_name = input("What the name of the file you want to create?")
+                        try:
+                            open(ask_name, "x")
+                            CommandSay(answer="DONE", color="OKGREEN")
+                        except FileExistsError:
+                            ask_del_create = input("This file already exist do you want to delete it. if yes type 'Y'")
+                            if ask_del_create == "Y" or ask_del_create == "y":
+                                os.remove(ask_name)
+                                CommandSay(answer="DONE", color="OKGREEN")
+                        except UnboundLocalError:
+                                pass
+                else:
+                    CommandSay(answer="This Function isn't available within this mode", color="FALI")
 
         if Command == "latest":
             LCommand = Command
@@ -172,9 +175,10 @@ try:
                 print("DONE")
 
         if Command == "gen password":
-            LCommand = Command
-            Boot.Run = True
-            Boot.SecondaryTask(file_name="Password_Gen", stay_end=True)
+            if not safe_md:
+                LCommand = Command
+                Boot.Run = True
+                Boot.SecondaryTask(file_name="Password_Gen", stay_end=True)
 
 
         if Command == "Exit":
@@ -183,39 +187,45 @@ try:
                 ask_exit = input("Are you sure. if yes press 'Y' and hit return")
                 if ask_exit == "Y" or ask_exit == "y":
                     sys.exit()
-            elif MD == "2" or MD == "9":
+            elif MD == "2" or MD == "9" or MD == "3":
                 sys.exit()
 
             
         if Command == "jump":
-            LCommand = Command
-            jump = True
+            if not safe_md:
+                LCommand = Command
+                jump = True
 
         if Command == "jump user":
-            LCommand = Command
-            jump_user = True
+            if not safe_md:
+                LCommand = Command
+                jump_user = True
 
 
         if Command == "print md":
-            CommandSay(answer=MD)
+            if not safe_md:
+                CommandSay(answer=MD)
+            else:
+                CommandSay("You Are in Safe-Mode")
 
 
 
         if Command == "talk":
-            LCommand = Command
-            if net:
-                ask_type = input("do you want to be host or reciever\nif you want to be host press 1 otherwise prees 2")
-                if ask_type == "1":
-                    Server.chat()
+            if not safe_md:
+                LCommand = Command
+                if net:
+                    ask_type = input("do you want to be host or reciever\nif you want to be host press 1 otherwise prees 2")
+                    if ask_type == "1":
+                        Server.chat()
+                    else:
+                        ask_recv = input("To Which IP you want to talk to\nType Below!\n:")
+                        CommandSay(answer=ask_recv)
+                        try: 
+                            client.Chat(IP=ask_recv)
+                        except ConnectionRefusedError:
+                            CommandSay(answer="This User is Unavilable at the moment\ntry again later", color="WARNING")
                 else:
-                    ask_recv = input("To Which IP you want to talk to\nType Below!\n:")
-                    CommandSay(answer=ask_recv)
-                    try: 
-                        client.Chat(IP=ask_recv)
-                    except ConnectionRefusedError:
-                        CommandSay(answer="This User is Unavilable at the moment\ntry again later", color="WARNING")
-            else:
-                CommandSay(answer="You Are in Safe Mode you can't connect to the internet right now")
+                    CommandSay(answer="You Are in Safe Mode you can't connect to the internet right now")
 
 
 
@@ -234,40 +244,44 @@ try:
                 os.system(f"more {ask_file}")
         
         if Command == "edit file":
-            if MD == "2":           
-                if cmd_pl == "1" or cmd_pl == "3":
-                    ask_file = input("type the name of the file you want to edit\n:")
-                    if ask_file.endswith(".py"):
-                        os.system(f"vim {ask_file}")
-                    else:
-                        os.system(f"nano {ask_file}")
-                elif cmd_pl == "2":
-                    CommandSay(answer="You can't edit files within The Windows Command Prompt", color="FAIL")
-            else:
-                LCommand = Command
-                CommandSay(answer="This Function isn't available within this mode", color="FALI")
+            if not safe_md:
+                if MD == "2":           
+                    if cmd_pl == "1" or cmd_pl == "3":
+                        ask_file = input("type the name of the file you want to edit\n:")
+                        if ask_file.endswith(".py"):
+                            os.system(f"vim {ask_file}")
+                        else:
+                            os.system(f"nano {ask_file}")
+                    elif cmd_pl == "2":
+                        CommandSay(answer="You can't edit files within The Windows Command Prompt", color="FAIL")
+                else:
+                    LCommand = Command
+                    CommandSay(answer="This Function isn't available within this mode", color="FALI")
         
         if Command == "weather forecast":
-            LCommand = Command
-            if net:
-                weather = os.system("curl wttr.in/")
-                CommandSay(answer="This is a fork from @igor_chubin", color="UNDERLINE") 
-            else:
-                CommandSay(answer="You Are in Safe Mode so you can't connect to the internet right now")
+            if not safe_md:
+                LCommand = Command
+                if net:
+                    weather = os.system("curl wttr.in/")
+                    CommandSay(answer="This is a fork from @igor_chubin", color="UNDERLINE") 
+                else:
+                    CommandSay(answer="You Are in Safe Mode so you can't connect to the internet right now")
 
         if Command == "activity monitor":
-            if MD == "2" or MD == "9":
-                if cmd_pl == "1" or cmd_pl == "3":
-                    os.system('top')
-            else:
-                LCommand = Command
-                CommandSay(answer="This Function isn't available within this mode", color="FALI")
+            if not safe_md:
+                if MD == "2" or MD == "9":
+                    if cmd_pl == "1" or cmd_pl == "3":
+                        os.system('top')
+                else:
+                    LCommand = Command
+                    CommandSay(answer="This Function isn't available within this mode", color="FALI")
 
         if Command == "countdown":
-            LCommand = Command
-            Boot.Run = True
-            Boot.SecondaryTask(file_name="countdown", stay_end=False)
-            
+            if not safe_md:
+                LCommand = Command
+                Boot.Run = True
+                Boot.SecondaryTask(file_name="countdown", stay_end=False)
+                
             
         
         if Command == "check site status":
@@ -281,16 +295,17 @@ try:
 
 
         if Command == "devices":
-            if not settings.pl == "2":
-                Boot.Run = True
-                if settings.server_use:
-                    import LocalNetworking.server
+            if not safe_md:
+                if not settings.pl == "2":
+                    Boot.Run = True
+                    if settings.server_use:
+                        import LocalNetworking.server
+                    else:
+                        import LocalNetworking.auth
+                        if LocalNetworking.auth.DONE:
+                            Boot.SecondaryTask(file_name="Handle-External-Devices", stay_end=True)
                 else:
-                    import LocalNetworking.auth
-                    if LocalNetworking.auth.DONE:
-                        Boot.SecondaryTask(file_name="Handle-External-Devices", stay_end=True)
-            else:
-                CommandSay("LocalNetworking Isn't Supported On Windown Yet\nIt's Under Development :)")
+                    CommandSay("LocalNetworking Isn't Supported On Windown Yet\nIt's Under Development :)")
 
     class bcolors:
         HEADER = '\033[95m'
